Breaking up contentView more
